requirements.txt
Flask
langchain
sentence-transformers
pypdf
python-dotenv
groq
langchain-community
langchain-huggingface
langchain-text-splitters
